db_migration_tool/requirements.txt
# Core dependencies
PySide6>=6.5.0
psycopg[binary]>=3.1.18
psycopg-pool>=3.2.0
psycopg2-binary>=2.9.9

# UI theme
qdarkstyle>=3.2.3

# Database
SQLAlchemy>=2.0.23

# Security
cryptography>=41.0.7

# Development
PyInstaller>=6.3.0

# Testing
pytest>=7.4.3
pytest-qt>=4.2.0